Moved synthesizer to pianoroll scripts and abstracted common playback code.
