+++ src/media/AudioDisplayComponent.h
@@ -41,13 +41,6 @@ public:
 
     void loadMediaFile(const URL& filePath) override;
 
-    void setPlaybackPosition(double t) override { transportSource.setPosition(t); }
-    double getPlaybackPosition() override { return transportSource.getCurrentPosition(); }
-
-    bool isPlaying() override { return transportSource.isPlaying(); }
-    void startPlaying() override { transportSource.start(); }
-    void stopPlaying() override { transportSource.stop(); }
-
     double getTotalLengthInSecs() override { return thumbnail.getTotalLength(); }
     double getTimeAtOrigin() override { visibleRange.getStart(); }
 
@@ -60,14 +53,8 @@ private:
     void postLoadActions(const URL& filePath) override;
 
     TimeSliceThread thread{ "Audio File Thread" };
-    
-    AudioFormatManager formatManager;
-    AudioDeviceManager deviceManager;
 
     std::unique_ptr<AudioFormatReaderSource> audioFileSource;
-
-    AudioSourcePlayer sourcePlayer;
-    AudioTransportSource transportSource;
 
     AudioThumbnailCache thumbnailCache{ 5 };
     AudioThumbnail thumbnail = AudioThumbnail(512, formatManager, thumbnailCache);
