+++ src/media/MidiDisplayComponent.h
@@ -1,9 +1,7 @@
 #pragma once
 
-#include <juce_audio_utils/juce_audio_utils.h>
-
 #include "../pianoroll/PianoRollComponent.hpp"
-#include "../synthesizer/SynthAudioSource.h"
+#include "../pianoroll/SynthAudioSource.h"
 #include "MediaDisplayComponent.h"
 
 
@@ -26,12 +24,7 @@ public:
 
     void loadMediaFile(const URL& filePath) override;
 
-    void setPlaybackPosition(double t) override { transportSource.setPosition(t); };
-    double getPlaybackPosition() override { transportSource.getCurrentPosition(); };
-
-    bool isPlaying() override { return transportSource.isPlaying(); };
     void startPlaying() override;
-    void stopPlaying() override { transportSource.stop(); };
 
     double getTotalLengthInSecs() override { return totalLengthInSecs; }
     float getPixelsPerSecond() override { return pianoRoll.getResolution(); }
@@ -48,13 +41,6 @@ private:
 
     double totalLengthInSecs;
 
-    TimeSliceThread thread{ "MIDI File Thread" };
-
-    AudioFormatManager formatManager;
-    AudioDeviceManager deviceManager;
-
-    AudioSourcePlayer sourcePlayer;
-    AudioTransportSource transportSource;
     SynthAudioSource synthAudioSource;
 
     PianoRollComponent pianoRoll{70, 5, scrollBarSize, controlSpacing};
