+++ src/pianoroll/SynthAudioSource.h
